Move Deliveries to child component of APp
